requirements.txt
streamlit
numpy
plotly
Pillow
deep-translator